Move caasp specific code to plugin
