+++ java/code/plugins/caasp/src/main/java/META-INF/MANIFEST.MF
