+++ java/code/plugins/caasp/src/main/java/com/suse/manager/caasp/CaaspPackageProfileUpdate.java
@@ -1,0 +1,65 @@
+/**
+ * Copyright (c) 2016 SUSE LLC
+ * <p>
+ * This software is licensed to you under the GNU General Public License,
+ * version 2 (GPLv2). There is NO WARRANTY for this software, express or
+ * implied, including the implied warranties of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
+ * along with this software; if not, see
+ * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
+ * <p>
+ * Red Hat trademarks are not licensed under GPLv2. No permission is
+ * granted to use or replicate Red Hat trademarks that are incorporated
+ * in this software or its documentation.
+ */
+package com.suse.manager.caasp;
+
+import com.redhat.rhn.common.messaging.MessageQueue;
+import com.redhat.rhn.domain.formula.FormulaFactory;
+import com.redhat.rhn.domain.server.MinionServer;
+import com.redhat.rhn.domain.server.Server;
+import com.redhat.rhn.frontend.events.RefreshPillarEvent;
+import com.redhat.rhn.manager.formula.FormulaManager;
+import com.suse.manager.extensions.PackageProfileUpdateExtensionPoint;
+import com.suse.manager.utils.SaltUtils;
+import org.apache.log4j.Logger;
+import org.pf4j.Extension;
+
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+@Extension
+public class CaaspPackageProfileUpdate implements PackageProfileUpdateExtensionPoint {
+
+    private static final Logger LOG = Logger.getLogger(CaaspPackageProfileUpdate.class);
+
+    public static final String CAASP_PRODUCT_IDENTIFIER = "caasp";
+
+    @Override
+    public void onProfileUpdate(Server server) {
+        // enable minion blackout (= locking) via pillar
+        if (server.asMinionServer().isPresent() && server.getInstalledProducts().stream()
+                .anyMatch(p -> p.getName().equalsIgnoreCase(CAASP_PRODUCT_IDENTIFIER))) {
+            // Minion blackout is only enabled for nodes that have installed the `caasp-*` package
+            Map<String, Object> data = new HashMap<>();
+            data.put("minion_blackout", true);
+            // List of Salt `module.function` that are allowed in blackout mode
+            data.put("minion_blackout_whitelist", Arrays.asList(
+                    "test.ping",
+                    "grains.item",
+                    "grains.items"
+            ));
+            try {
+                FormulaManager.getInstance().enableFormula(server.getMinionId(), SaltUtils.SYSTEM_LOCK_FORMULA);
+                FormulaFactory.saveServerFormulaData(data, server.getMinionId(), SaltUtils.SYSTEM_LOCK_FORMULA);
+                MessageQueue.publish(new RefreshPillarEvent());
+            }
+            catch (IOException e) {
+                LOG.error("Could not enable blackout formula", e);
+            }
+        }
+    }
+
+}
