+++ java/code/plugins/caasp/src/test/resources/log4j.properties
@@ -1,0 +1,12 @@
+log4j.rootLogger=DEBUG,CONSOLE                                                                                                                                                  
+
+# Hibernate logging options (INFO only shows startup messages)
+log4j.logger.org.hibernate=DEBUG                                                                                                                                            
+
+# Log JDBC bind parameter runtime arguments
+log4j.logger.org.hibernate.type=INFO
+
+log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
+log4j.appender.CONSOLE.Target=System.out
+log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
+log4j.appender.CONSOLE.layout.conversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} %-5p - %m%n
