+++ java/code/src/com/suse/manager/utils/SaltUtils.java
@@ -176,7 +176,7 @@ public class SaltUtils {
 
     public static final SaltUtils INSTANCE = new SaltUtils();
 
-    public static final String CAASP_PRODUCT_IDENTIFIER = "caasp";
+    // TODO move this constant somewhere else
     public static final String SYSTEM_LOCK_FORMULA = "system-lock";
 
     private Path scriptsDir = Paths.get(SUMA_STATE_FILES_ROOT_PATH, SCRIPTS_DIR);
@@ -1231,31 +1231,6 @@ public class SaltUtils {
         Plugins.instance().getExtensions(PackageProfileUpdateExtensionPoint.class)
                 .forEach(listener -> listener.onProfileUpdate(server));
 
-        // For special nodes: enable minion blackout (= locking) via pillar
-        enableMinionSystemLockForSpecialNodes(server);
-    }
-
-    private static void enableMinionSystemLockForSpecialNodes(MinionServer server) {
-        if (server.getInstalledProducts().stream()
-                .anyMatch(p -> p.getName().equalsIgnoreCase(CAASP_PRODUCT_IDENTIFIER))) {
-            // Minion blackout is only enabled for nodes that have installed the `caasp-*` package
-            Map<String, Object> data = new HashMap<>();
-            data.put("minion_blackout", true);
-            // List of Salt `module.function` that are allowed in blackout mode
-            data.put("minion_blackout_whitelist", Arrays.asList(
-                    "test.ping",
-                    "grains.item",
-                    "grains.items"
-            ));
-            try {
-                FormulaManager.getInstance().enableFormula(server.getMinionId(), SYSTEM_LOCK_FORMULA);
-                FormulaFactory.saveServerFormulaData(data, server.getMinionId(), SYSTEM_LOCK_FORMULA);
-                MessageQueue.publish(new RefreshPillarEvent());
-            }
-            catch (IOException e) {
-                LOG.error("Could not enable blackout formula", e);
-            }
-        }
     }
 
     /**
