+++ java/code/src/log4j.properties
@@ -4,7 +4,7 @@ log4j.appender.RootAppender.MaxFileSize=10MB
 log4j.appender.RootAppender.MaxBackupIndex=5
 log4j.appender.RootAppender.layout=org.apache.log4j.PatternLayout
 log4j.appender.RootAppender.layout.ConversionPattern=%d [%t] %-5p %c - %m%n
-log4j.rootLogger=WARN,RootAppender
+log4j.rootLogger=INFO,RootAppender
 
 ## XmlRpc Logging
